nilfs-utils: add support for NILFS_IOCTL_SET_SUINFO ioctl
With this ioctl the segment usage information in the SUFILE can be
updated from userspace.

This is useful, because it allows the GC to modify and update segment
usage entries for specific segments, which enables it to avoid
unnecessary write operations.

If a segment needs to be cleaned, but there is no or very little free
space to be gained, the cleaning operation basically degrades to
needless expensive copying of data. In the end the only thing that
changes is the location of the data and a timestamp in the segment
usage info. With this ioctl the GC can skip the copying and update the
segment usage entries directly instead.

Additionally this patch implements a simple check in
nilfs_reclaim_segment. If the number of free blocks that can be gained
by cleaning the segments is below the threshold of minblocks, it simply
updates the segment usage information instead.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>