nilfs-utils: remove reliance on sui_nblocks to read segment
Since sui_nblocks is reused to represent the number of live blocks in a
segment it cannot be used any more to mark the end of the segment.
Instead the sequence number of the partial segments is checked. The
sequence number of partial segments should all be the same. The usual
CRC checks should be enough to reliably determine the end of a segment.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>