nilfs-utils: add extra flags to nilfs_vdesc and update sui_nblocks
This patch adds extra flags to nilfs_vdesc that indicate the reason for
which a particular block is considered alive. If it is because of a
snapshot, then the snapshot flag is set, if it is because of the
protection period, then that flag is set.

This information is useful to determine the number of live blocks in a
segment. If a block is part of a snapshot, it is counted as alive, if it
is alive because of the protection period it is counted as reclaimable.
These flags are used both in userspace and by the kernel.

Additionally this patch adds code that calculates the correct number of
live blocks per segment if nilfs_set_suinfo() is used.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>