nilfs-utils: add support for NILFS_FEATURE_COMPAT_TRACK_SNAPSHOTS
This patch adds support for the NILFS_FEATURE_COMPAT_TRACK_SNAPSHOTS
feature. It depends both on the NILFS_FEATURE_COMPAT_SUFILE_EXTENSION
and on NILFS_FEATURE_COMPAT_TRACK_LIVE_BLKS to work. It can be enabled
or disabled at file system creation time with mkfs or later with
nilfs-tune.

If this feature is activated, the creation and deletion of snapshots
will be more expensive, but the normal file system operation is not
affected at all. Therefore, depending on the intended usage pattern
for the file system, it can be beneficial to either activate or
deactivate this feature. It is not essential for the GC operation, but
it will improve the GC performance significantly, at the cost of
longer snapshot creation and deletion times.

Additionally this patch adds the NILFS_IOCTL_SET_INC_FLAGS ioctl to
correctly update the flags in the DAT file if the set_suinfo ioctl is
used.

For the tracking of snapshots it is necessary to update the de_ss
field of the DAT entries to agree with the corresponding su_nlive_blks
counter in the SUFILE. This basically prevents the counter from being
incremented multiple times for one and the same block. Normally this
happens when the GC moves the live blocks to a new segment. However if
set_suinfo ioctl is used, this cannot happen. Therefore  the new ioctl
is needed.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>